refactor(db): use correct virtual types for tests
