+++ packages/db/test/unit/reset-queries.test.ts
@@ -4,7 +4,7 @@ import { getMigrationQueries } from '../../dist/core/cli/migration-queries.js';
 import { MIGRATION_VERSION } from '../../dist/core/consts.js';
 import { tableSchema } from '../../dist/core/schemas.js';
 import type { ResolvedDBTable } from '../../dist/core/types.js';
-import { column, defineTable } from '../../dist/runtime/virtual.js';
+import { column, defineTable } from './virtual.ts';
 import { asResolved } from '../test-utils.ts';
 
 const TABLE_NAME = 'Users';
